+++ dascore/utils/patch.py
@@ -595,8 +595,8 @@ def get_dim_axis_value(
 
     Examples
     --------
-    >>> import dascore as dc
-    >>> from dascore.utils.patch import get_dim_axis_value
+    import dascore.proc.coord    >>> import dascore as dc
+    import dascore.proc.coords    >>> from dascore.utils.patch import get_dim_axis_value
     >>>
     >>> patch = dc.get_example_patch()
     >>>
