+++ tests/test_proc/test_detrend.py
@@ -13,5 +13,5 @@ class TestDetrend:
         new = random_patch.new(data=random_patch.data + 10)
         # perform detrend, ensure all mean values are close to zero
         det = new.detrend(dim="time", type="linear")
-        means = np.mean(det.data, axis=det.get_axis("time"))
+        means = np.mean(det.data, axis=random_patch.get_axis("time"))
         assert np.allclose(means, 0)
